requirements.txt
accelerate==0.31.0
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
bitarray==2.9.2
blinker==1.8.2
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
colorama==0.4.6
comm==0.2.2
comtypes==1.4.8
confection==0.1.5
contourpy==1.2.1
cramjam==2.9.0
cryptography==42.0.7
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.6
datasets==2.19.2
debugpy==1.8.1
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.8
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
Django==5.1.2
dnspython==2.7.0
env==0.1.0
et-xmlfile==1.1.0
evaluate==0.4.2
eventkit==1.0.3
executing==2.0.1
fastapi==0.115.3
fastparquet==2024.5.0
ffmpy==0.4.0
filelock==3.14.0
Flask==3.0.3
fonttools==4.53.1
fpdf==1.7.2
frozendict==2.4.4
frozenlist==1.4.1
fsspec==2024.3.1
gradio==5.3.0
gradio_client==1.4.2
greenlet==3.0.3
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.26.1
ib-insync==0.9.86
idna==3.7
importlib_resources==6.4.5
intel-openmp==2021.4.0
ipykernel==6.29.4
ipython==8.24.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
langchain==0.2.7
langchain-community==0.2.6
langchain-core==0.2.12
langchain-openai==0.1.14
langchain-text-splitters==0.2.0
langcodes==3.4.0
langsmith==0.1.84
language_data==1.2.0
lightgbm==4.3.0
limits==3.13.0
llama-cloud==0.1.4
llama-index==0.11.20
llama-index-agent-openai==0.3.4
llama-index-cli==0.3.1
llama-index-core
llama-index-embeddings-huggingface==0.2.1
llama-index-embeddings-openai==0.2.5
llama-index-extractors-entity==0.1.2
llama-index-indices-managed-llama-cloud==0.4.0
llama-index-legacy==0.9.48
llama-index-llms-llama-cpp==0.1.3
llama-index-llms-openai==0.2.16
llama-index-multi-modal-llms-openai==0.2.3
llama-index-program-openai==0.2.0
llama-index-question-gen-openai==0.2.0
llama-index-readers-file==0.2.2
llama-index-readers-llama-parse==0.3.0
llama-parse==0.5.11
llama_cpp_python==0.2.77
llamaindex-py-client==0.1.19
llmsherpa==0.1.4
lxml==5.2.2
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
minijinja==2.0.1
mkl==2021.4.0
motor==3.6.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
multitasking==0.0.11
murmurhash==1.0.10
mypy-extensions==1.0.0
mysql-connector-python==9.1.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
numpy==1.26.4
openai==1.52.1
opencv-python==4.10.0.84
openpyxl==3.1.4
orjson==3.10.3
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.2
parso==0.8.4
pdfminer.six==20231228
pdfplumber==0.11.0
pdfreader==0.1.15
peewee==3.17.5
pillow==10.3.0
platformdirs==4.2.2
playwright==1.44.0
preshed==3.0.9
prompt_toolkit==3.0.44
protobuf==5.27.2
psutil==5.9.8
pure-eval==0.2.2
pyarrow==16.1.0
pyarrow-hotfix==0.6
PyAudio==0.2.14
pycparser==2.22
pycryptodome==3.20.0
pydantic
pydantic_core==2.18.3
pydub==0.25.1
pyee==11.1.0
Pygments==2.18.0
pymongo==4.9.2
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
pyparsing==3.1.2
pypdf==4.2.0
PyPDF2==3.0.1
pypdfium2==4.30.0
pypiwin32==223
pysbd==0.3.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
pyttsx3==2.98
pytz==2024.1
# pywin32
PyYAML==6.0.1
pyzmq==26.0.3
ragas==0.1.10
regex==2024.5.15
requests==2.32.2
rich==13.7.1
ruff==0.7.0
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
selenium==4.22.0
semantic-version==2.10.0
sentence-transformers==2.7.0
seqeval==1.2.2
setuptools==70.0.0
shellingham==1.5.4
six==1.16.0
slowapi==0.1.9
smart-open==7.0.4
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
span-marker==1.5.0
SpeechRecognition==3.11.0
SQLAlchemy==2.0.30
sqlparse==0.5.1
srsly==2.4.8
stack-data==0.6.3
starlette==0.41.0
striprtf==0.0.26
sympy==1.12.1
tbb==2021.12.0
tenacity==8.3.0
thinc==8.2.4
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
toml==0.10.2
tomlkit==0.12.0
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
transformers==4.41.2
trio==0.25.1
trio-websocket==0.11.1
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.1
tzdata==2024.1
upstox-python-sdk==2.4.1
urllib3==2.2.1
uuid==1.30
uvicorn==0.32.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.4
wrapt==1.16.0
wsproto==1.2.0
xxhash==3.4.1
yarl==1.9.4
yfinance==0.2.40